consumer/requirements.txt
kafka-python==2.0.2
minio==7.1.15
python-dotenv==1.0.0
pandas==2.0.3
pyarrow==14.0.1
ujson==5.8.0
tenacity==8.2.3
prometheus-client==0.17.1